# internal/apidocs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Velora API",
    "version": "0.6.0",
    "description": "Dating and social API for the Velora clients.\n\n**Getting a token.** Run `POST /auth/register` with a throwaway address, copy `accessToken` from the response, press **Authorize**, and paste it. Access tokens last 15 minutes; `POST /auth/refresh` rotates them, and spending one refresh token twice revokes the whole family.\n\n**Errors** are RFC 9457 problem documents: `type`, `title`, `status`, `detail`, and `instance` carrying the request id, so an error a person saw can be traced to a log line.\n\n**Blocking** is enforced in every read path and in both directions. A blocked profile answers 404 exactly like a deleted one — which of the two happened is deliberately not disclosed."
  },
  "servers": [
    { "url": "/api", "description": "Through nginx on the test server" },
    { "url": "/", "description": "Directly against the API" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The `accessToken` from register, login, or refresh."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457. Every error the API returns uses this shape.",
        "properties": {
          "type": { "type": "string", "example": "about:blank#bad-request" },
          "title": { "type": "string", "example": "Bad request" },
          "status": { "type": "integer", "example": 400 },
          "detail": { "type": "string", "example": "Your first name cannot be empty." },
          "instance": {
            "type": "string",
            "description": "The request id, for finding this in the logs.",
            "example": "c3588ff1957f8aa6"
          }
        }
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "displayName": { "type": "string", "example": "Sam" },
          "email": { "type": "string", "example": "sam@velora.test" },
          "phone": { "type": "string" },
          "method": { "type": "string", "enum": ["email", "phone", "google", "apple"] },
          "isPremium": { "type": "boolean" }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/AuthUser" },
          "accessToken": { "type": "string", "description": "Bearer token, 15 minutes." },
          "refreshToken": {
            "type": "string",
            "description": "Rotates on use. Presenting a spent one revokes the family."
          },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Understood and refused. `detail` is safe to show a person.",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Create an account",
        "description": "Start here. Copy `accessToken` from the response into **Authorize**.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "you+1@velora.test" },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "disposable-test-pw"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Session" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
  }
}